MapView의 NavigationBar를 disable하기 위해 modifire추가 #17
