+++ Scope/UserInterface/Sources/Views/Home/Action/HomeAction.swift
@@ -1,0 +1,11 @@
+//
+//  HomeAction.swift
+//  UserInterface
+//
+//  Created by 조요한 on 2021/12/29.
+//  Copyright © 2021 com.devyhan93.scope. All rights reserved.
+//
+
+public enum HomeAction {
+  case updateRegion(CoordinateRegion?)
+}
